Minor Changes to Usages
1. Changed total duration to using seconds instead of using minutes for data accuracy
2. Changed duration and value to be double instead of float.
3. Added Formatting to Line Chart to display y axis as dd:hh:mm:ss

+++ 8.1.0/aspnet-core/src/SYT.BnBCheckIn.Application/Usages/Dto/DayUsage.cs
@@ -7,7 +7,7 @@ namespace SYT.BnBCheckIn.Usages.Dto
 	public class DayUsage
 	{
 		public string name { get; set; }
-		public float value { get; set; }
+		public double value { get; set; }
 		public string start { get; set; }
 		public string end { get; set; }
     }
@@ -21,7 +21,7 @@ namespace SYT.BnBCheckIn.Usages.Dto
 		public string Building { get; set; }
 		public DateTime StartTime { get; set; }
 		public DateTime EndTime { get; set; }
-		public float Duration { get; set; }
+		public double Duration { get; set; }
 	}
 
 	public class DayUsageWithUnits
